Self Review: Moved inputs components into a folder
