+++ app/(Components)/(Form)/Inputs/InputDate.tsx
@@ -2,12 +2,17 @@ import React from 'react';
 import DatePicker, {registerLocale} from 'react-datepicker';
 import fr from 'date-fns/locale/fr';
 import 'react-datepicker/dist/react-datepicker.css';
-import {InputDateProps} from './InputDate';
-import styles from './styles.module.css';
+import styles from '../styles.module.css';
 
 registerLocale('fr', fr);
 
-export const InputTime: React.FC<InputDateProps> = ({
+export interface InputDateProps {
+	value: Date;
+	setValue: (date: Date) => void;
+	label: string;
+}
+
+export const InputDate: React.FC<InputDateProps> = ({
 	value,
 	setValue,
 	label,
@@ -16,14 +21,9 @@ export const InputTime: React.FC<InputDateProps> = ({
 		<label className={styles.label}>
 			{label}
 			<DatePicker
-				showTimeSelect
-				showTimeSelectOnly
-				timeIntervals={15}
-				dateFormat="HH:MM"
-				locale="fr"
 				className={styles.input}
+				locale="fr"
 				selected={value}
-				timeCaption="Heure"
 				onChange={(date: Date) => setValue(date)}
 			/>
 		</label>
