+++ app/(Components)/(Form)/Inputs/InputTime.tsx
@@ -2,17 +2,12 @@ import React from 'react';
 import DatePicker, {registerLocale} from 'react-datepicker';
 import fr from 'date-fns/locale/fr';
 import 'react-datepicker/dist/react-datepicker.css';
-import styles from './styles.module.css';
+import {InputDateProps} from './InputDate';
+import styles from '../styles.module.css';
 
 registerLocale('fr', fr);
 
-export interface InputDateProps {
-	value: Date;
-	setValue: (date: Date) => void;
-	label: string;
-}
-
-export const InputDate: React.FC<InputDateProps> = ({
+export const InputTime: React.FC<InputDateProps> = ({
 	value,
 	setValue,
 	label,
@@ -21,9 +16,14 @@ export const InputDate: React.FC<InputDateProps> = ({
 		<label className={styles.label}>
 			{label}
 			<DatePicker
-				className={styles.input}
+				showTimeSelect
+				showTimeSelectOnly
+				timeIntervals={15}
+				dateFormat="HH:MM"
 				locale="fr"
+				className={styles.input}
 				selected={value}
+				timeCaption="Heure"
 				onChange={(date: Date) => setValue(date)}
 			/>
 		</label>
